using 32 bit FNVHash instead of murmur, pass for co-sim

+++ project/Server/LZW/fpga/lzw_kernel.h
@@ -26,7 +26,7 @@
 #define BUCKET_SIZE	3
 
 typedef struct hashmap_entry {
-  uint64_t key;
+  uint32_t key;
   uint32_t code;
 } hashmap_entry_t;
